Increased largePageDataBytes threshold
Before this change, we'd get lots of warnings about the data in
`getStaticProps` being too large
(https://nextjs.org/docs/messages/large-page-data). After thinking about
it a bit: documentation for a module is sometimes going to be large,
that's life. If / once we refactor the page structure this may get
better, but even then, documentation for a _class_ can be big. And
that's OK.

[ghstack-poisoned]